# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gestao-financeira_backend_internal_account_application_dtos.AccountOutput:
    properties:
      account_id:
        type: string
      balance:
        type: number
      context:
        type: string
      created_at:
        type: string
      currency:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_account_application_dtos.CreateAccountInput:
    properties:
      context:
        enum:
        - PERSONAL
        - BUSINESS
        type: string
      currency:
        enum:
        - BRL
        - USD
        - EUR
        type: string
      initial_balance:
        minimum: 0
        type: number
      name:
        maxLength: 100
        minLength: 3
        type: string
      type:
        enum:
        - BANK
        - WALLET
        - INVESTMENT
        - CREDIT_CARD
        type: string
      user_id:
        type: string
    required:
    - context
    - currency
    - name
    - type
    - user_id
    type: object
  gestao-financeira_backend_internal_account_application_dtos.CreateAccountOutput:
    properties:
      account_id:
        type: string
      balance:
        type: number
      context:
        type: string
      created_at:
        type: string
      currency:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_account_application_dtos.GetAccountOutput:
    properties:
      account_id:
        type: string
      balance:
        type: number
      context:
        type: string
      created_at:
        type: string
      currency:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_account_application_dtos.ListAccountsOutput:
    properties:
      accounts:
        items:
          $ref: '#/definitions/gestao-financeira_backend_internal_account_application_dtos.AccountOutput'
        type: array
      count:
        type: integer
    type: object
  gestao-financeira_backend_internal_identity_application_dtos.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  gestao-financeira_backend_internal_identity_application_dtos.LoginOutput:
    properties:
      email:
        type: string
      expires_in:
        description: Token expiration in seconds
        type: integer
      first_name:
        type: string
      full_name:
        type: string
      last_name:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_identity_application_dtos.RegisterUserInput:
    description: User registration input with email, password, first name and last
      name
    properties:
      email:
        description: Email address of the user (must be a valid email format)
        example: user@example.com
        type: string
      first_name:
        description: First name of the user (minimum 2 characters)
        example: John
        minLength: 2
        type: string
      last_name:
        description: Last name of the user (minimum 2 characters)
        example: Doe
        minLength: 2
        type: string
      password:
        description: Password for the user account (minimum 8 characters)
        example: SecurePass123
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gestaofinanceira.com
    name: API Support
  description: API REST para gestão financeira pessoal e profissional. Gerencia usuários,
    contas e transações financeiras.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gestão Financeira API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Lists all accounts for the authenticated user. Optionally filter
        by context (PERSONAL or BUSINESS).
      parameters:
      - description: Filter by context (PERSONAL or BUSINESS)
        in: query
        name: context
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of accounts with count
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_account_application_dtos.ListAccountsOutput'
        "400":
          description: Bad request - invalid user ID or context
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new account for the authenticated user. Supports BANK,
        WALLET, INVESTMENT, and CREDIT_CARD account types.
      parameters:
      - description: Account creation data (name, type, initial_balance, currency,
          context)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gestao-financeira_backend_internal_account_application_dtos.CreateAccountInput'
      produces:
      - application/json
      responses:
        "201":
          description: Account data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_account_application_dtos.CreateAccountOutput'
        "400":
          description: Bad request - invalid input data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict - account already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific account by its ID. Only returns accounts that
        belong to the authenticated user.
      parameters:
      - description: Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_account_application_dtos.GetAccountOutput'
        "400":
          description: Bad request - invalid account ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - account does not belong to user
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - account does not exist
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get account by ID
      tags:
      - accounts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, returns a JWT token
        for API access
      parameters:
      - description: User login credentials (email and password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gestao-financeira_backend_internal_identity_application_dtos.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token and user data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_identity_application_dtos.LoginOutput'
        "400":
          description: Bad request - invalid input data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid email or password
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - user account is inactive
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided email, password, first
        name, and last name.
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gestao-financeira_backend_internal_identity_application_dtos.RegisterUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: 'User registered successfully" "Example: {\"message\": \"User
            registered successfully\", \"data\": {\"user_id\": \"uuid\", \"email\":
            \"user@example.com\", \"first_name\": \"John\", \"last_name\": \"Doe\",
            \"full_name\": \"John Doe\"}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid input data (e.g., invalid email format,
            password too short, missing fields)
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict - User with this email already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error - An unexpected error occurred
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"